[src][scheduler] Split scheduler.c
